+++ base/deprecated.jl
@@ -1592,6 +1592,8 @@ end
 
 @deprecate_binding AbstractIOBuffer GenericIOBuffer false
 
+@deprecate String(io::GenericIOBuffer) String(take!(copy(io)))
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations